Conflicts:
	InscricoesCrescer/InscricoesCrescer/Controllers/AdministrativoController.cs

+++ InscricoesCrescer/InscricoesCrescer.Dominio/Candidato/ICandidatoRepositorio.cs
@@ -13,7 +13,7 @@ namespace InscricoesCrescer.Dominio.Candidato
 
         void Editar(CandidatoEntidade candidato);
 
-        CandidatoEntidade BuscarPorId(int id);
+        CandidatoEntidade BuscarPorId(long id);
 
         List<CandidatoEntidade> BuscarTodos();
